requirements.txt
alphasign
pyserial
